Separate header file for source and convert stages
