+++ morpheus/_lib/doca/CMakeLists.txt
@@ -24,12 +24,12 @@ add_library(morpheus_doca
   # Keep these sorted!
   src/doca_context.cpp
   src/doca_convert_kernel.cu
-  src/doca_convert.cpp
+  src/doca_convert_stage.cpp
   src/doca_rx_pipe.cpp
   src/doca_rx_queue.cpp
   src/doca_semaphore.cpp
   src/doca_source_kernel.cu
-  src/doca_source.cpp
+  src/doca_source_stage.cpp
   src/packet_data_buffer.cpp
   src/rte_context.cpp
 )
